Counting unanwered questions...
What strikes me as odd is that running jekyll seems to obliterate the
file 'custom.js' in the root of the directory. It is as if the
generation process copies it out of the _site directory before
proceeding.

I'm not sure what's going on, but I do need to figure that out, just
from a compilation/build process perspective.

Either way, this lets me add a cute button that shows how many
unanswered questions there are, and it takes the viewer directly to the
issues page for the course.