Adding in Bootstrap.
I want to be able to update color and the like. Perhaps including the
Bootstrap source is the wrong way to do that. At the same time, I want
the project to be self-contained.

I read about Git submodules, and they looked to be the difficult way to
do this. Ultimately, I'll get a Makefile going that rebuilds the
bootstrap.css file and puts it in place, so that modifications are easy.