New nav generation, pages...
There's now some Liquid templating intelligence that generates the
navbar. Specifically, it uses a list of categories in the _config.yaml
to decide what pages should be under which dropdowns. The result?

1. The page maintainer only needs a list of categories, which become
top-level headings in the navbar.

2. The pages that you want to include only need to be tagged with a

group: "Admin"

or similar to show up in the "Admin" group. This doesn't provide
ordering in the drop-down, but it makes it easy to add pages. It uses
the URL to the page such that it doesn't matter where in the site that
it is... the page should "just link" from the menu.